+++ src/BioSequences.jl
@@ -1,5 +1,5 @@
 # BioSequences.jl
-# =======
+# ===============
 #
 # A julia package for the representation and manipulation of biological sequences.
 #
@@ -10,6 +10,9 @@ module BioSequences
 
 export
     # Symbols
+    # -------
+    
+    # Types & aliases
     NucleicAcid,
     DNA,
     RNA,
@@ -49,13 +52,6 @@ export
     RNA_Gap,
     ACGU,
     ACGUN,
-    isGC,
-    iscompatible,
-    isambiguous,
-    iscertain,
-    isgap,
-    ispurine,
-    ispyrimidine,
     AminoAcid,
     AA_A,
     AA_R,
@@ -85,15 +81,79 @@ export
     AA_X,
     AA_Term,
     AA_Gap,
-
+    
+    # Predicates
+    isGC,
+    iscompatible,
+    isambiguous,
+    iscertain,
+    isgap,
+    ispurine,
+    ispyrimidine,
+    
+    
+    # Alphabets
+    # ---------
+    
+    # Types & aliases
+    Alphabet,
+    NucleicAcidAlphabet,
+    DNAAlphabet,
+    RNAAlphabet,
+    AminoAcidAlphabet,
+    CharAlphabet,
+    
+    
     # BioSequences
+    # ------------
+    
+    # Type & aliases
     BioSequence,
     NucleotideSeq,
+    
+    # Indexing
+    unsafe_setindex!,
+    
+    # Predicates
+    ispalindromic,
+    hasambiguity,
+    isrepetitive,
+    
+    
+    # LongSequence
+    # ------------
+    
+    # Type & aliases
     LongSequence,
     DNASequence,
     RNASequence,
     AminoAcidSequence,
     CharSequence,
+    
+    
+    # Skipmers
+    # --------
+    
+    # Type & aliases
+    Skipmer,
+    DNASkipmer,
+    RNASkipmer,
+    BigDNASkipmer,
+    BigRNASkipmer,
+    Kmer,
+    DNAKmer,
+    RNAKmer,
+    BigDNAKmer,
+    BigRNAKmer,
+    DNACodon,
+    RNACodon,
+    
+    # Transformations
+    canonical,
+    
+    # Iteration
+    neighbors,
+    
     seqname,
     hasseqname,
     sequence,
@@ -112,9 +172,7 @@ export
     symbols,
     gap,
     mismatches,
-    ispalindromic,
-    hasambiguity,
-    isrepetitive,
+    
     ambiguous_positions,
     gc_content,
     SamplerUniform,
@@ -123,27 +181,15 @@ export
     randdnaseq,
     randrnaseq,
     randaaseq,
-    canonical,
-    neighbors,
-    eachkmer,
     each,
+    eachcanonical,
     Composition,
     composition,
     NucleicAcidCounts,
-    Skipmer,
-    Kmer,
-    DNAKmer,
-    RNAKmer,
-    BigDNAKmer,
-    BigRNAKmer,
-    DNACodon,
-    RNACodon,
+    
     translate,
     ncbi_trans_table,
     ABIF,
-    
-    # Indexing
-    unsafe_setindex!,
     
     # Transformations
     complement,
@@ -152,15 +198,8 @@ export
     reverse_complement!,
     ungap,
     ungap!,
-
-    # Alphabets
-    Alphabet,
-    NucleicAcidAlphabet,
-    DNAAlphabet,
-    RNAAlphabet,
-    AminoAcidAlphabet,
-    CharAlphabet,
-
+    
+    
     # search
     ExactSearchQuery,
     ApproximateSearchQuery,
@@ -172,7 +211,7 @@ export
     PWM,
     maxscore,
     scoreat,
-
+    
     ReferenceSequence,
     Demultiplexer,
     demultiplex,
@@ -211,19 +250,33 @@ using Random
 
 BioSymbols.gap(::Type{Char}) = '-'
 
-include("trait_definitions/alphabet.jl")
+include("alphabet.jl")
+
+# Load the bit-twiddling internals that optimised BioSequences methods depend on.
 include("bit-manipulation/bit-manipulation.jl")
-include("biosequences/biosequence.jl")
+
+# The generic, abstract BioSequence type
+include("biosequence/biosequence.jl")
+
+# The definition of the LongSequence concrete type, and its method overloads...
 include("longsequences/longsequence.jl")
 include("longsequences/hash.jl")
 include("longsequences/randseq.jl")
-#include("shortsequence/shortsequence.jl")
+
+# The definition of the Skipmer concrete type, and its method overloads...
 include("skipmers/skipmer.jl")
+
+# The definition of the ReferenceSequence concrete type, and its method overloads...
 include("nmask.jl")
 include("refseq/refseq.jl")
+
+# The generic iterators for any BioSequence...
 include("iterators/ambiguous.jl")
 include("iterators/eachkmer.jl")
+include("iterators/eachskipmer.jl")
+
 include("composition.jl")
+
 include("geneticcode.jl")
 include("demultiplexer.jl")