+++ src/biosequence/operations.jl
@@ -21,6 +21,36 @@ function Base.count(f::Function, seq::BioSequence)
     return n
 end
 
+
+# Finders
+# -------
+
+function Base.findnext(val, seq::BioSequence, start::Integer)
+    checkbounds(seq, start)
+    v = convert(eltype(seq), val)
+    for i in start:lastindex(seq)
+        if inbounds_getindex(seq, i) == v
+            return i
+        end
+    end
+    return nothing
+end
+
+function Base.findprev(val, seq::BioSequence, start::Integer)
+    checkbounds(seq, start)
+    v = convert(eltype(seq), val)
+    for i in start:-1:1
+        if inbounds_getindex(seq, i) == v
+            return i
+        end
+    end
+    return nothing
+end
+
+Base.findfirst(val, seq::BioSequence) = findnext(val, seq, 1)
+Base.findlast(val, seq::BioSequence) = findprev(val, seq, lastindex(seq))
+
+
 # GC content
 # ----------