+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - All random `LongSequence` generator methods take an optional AbstractRNG argument.
 - Add methods to `randseq` to optimize random generation of `NucleicAcid` or `AminoAcid` `LongSequence`s.
 - BioGenerics is now a dependency - replaces BioCore.
+- A `CanonicalSkipmers` iterator that allows iteration over the Skipmers in a 
+  nucleotide sequence.  
 
 ### Changed
 - The abstract `Sequence` type is now called `BioSequence{A}`.